controler and tests work

+++ dndserver/src/main/java/com/gmail/phoenixjoe/dndserver/repository/ItemRepository.java
@@ -8,7 +8,7 @@ import com.gmail.phoenixjoe.dndserver.model.Item;
 
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
     
 
 }
